+++ includes/llms.template.hooks.php
@@ -36,6 +36,7 @@ add_action( 'lifterlms_single_course_after_summary', 'lifterlms_template_single_
 add_action( 'lifterlms_single_course_after_summary', 'lifterlms_template_pricing_table',             60 );
 add_action( 'lifterlms_single_course_after_summary', 'lifterlms_template_single_course_progress',    60 );
 add_action( 'lifterlms_single_course_after_summary', 'lifterlms_template_single_syllabus',           90 );
+add_action( 'lifterlms_single_course_after_summary', 'lifterlms_template_single_reviews',            100 );