+++ lifterlms.php
@@ -3,7 +3,7 @@
 * Plugin Name: LifterLMS
 * Plugin URI: https://lifterlms.com/
 * Description: LifterLMS, the #1 WordPress LMS solution, makes it easy to create, sell, and protect engaging online courses.
-* Version: 3.1.2
+* Version: 3.1.3
 * Author: Thomas Patrick Levy, codeBOX LLC, Mark Nelson
 * Author URI: http://gocodebox.com
 * Text Domain: lifterlms
@@ -35,7 +35,7 @@ require_once 'vendor/autoload.php';
  */
 final class LifterLMS {
 
-	public $version = '3.1.2';
+	public $version = '3.1.3';
 
 	protected static $_instance = null;
 
@@ -291,6 +291,8 @@ final class LifterLMS {
 			include_once( 'includes/shortcodes/class.llms.shortcode.checkout.php' );
 
 		}
+
+		require_once 'includes/class.llms.playnice.php';
 
 	}