+++ templates/course/outline-list-small.php
@@ -5,6 +5,8 @@
 ?>
 <div class="llms-widget-syllabus<?php echo ( $collapse ) ? ' llms-widget-syllabus--collapsible' : ''; ?>">
 
+	<?php do_action( 'lifterlms_outline_before' ); ?>
+
 	<ul class="llms-course-outline">
 
 		<?php //get section data
@@ -13,6 +15,8 @@
 			<li class="llms-section<?php echo ( $collapse ) ? ( $current_section && $section['id'] == $current_section ) ? ' llms-section--opened' : ' llms-section--closed' : ''; ?>">
 
 				<div class="section-header">
+
+					<?php do_action( 'lifterlms_outline_before_header' ); ?>
 
 					<?php if ( $collapse ) : ?>
 
@@ -26,6 +30,8 @@
 					<?php endif; ?>
 
 					<span class="section-title"><?php echo $section['title']; ?></span>
+
+					<?php do_action( 'lifterlms_outline_after_header' ); ?>
 
 				</div>
 
@@ -44,6 +50,8 @@
 
 								</span>
 
+								<?php do_action( 'lifterlms_outline_before_lesson_title', $lesson ); ?>
+
 								<span class="lesson-title <?php echo ( $lesson['is_complete'] ? 'done' : '' ); ?>">
 
 									<?php $l = new LLMS_Lesson( $lesson['id'] ); ?>
@@ -59,6 +67,8 @@
 									endif; ?>
 
 								</span>
+
+								<?php do_action( 'lifterlms_outline_after_lesson_title', $lesson ); ?>
 
 							</li>
 
@@ -76,9 +86,13 @@
 
 			<li class="llms-section llms-syllabus-footer">
 
+				<?php do_action( 'lifterlms_outline_before_footer' ); ?>
+
 				<a class="llms-button-text llms-collapse-toggle" data-action="open" href="#"><?php _e( 'Open All', 'lifterlms' ); ?></a>
 				<span>&middot;</span>
 				<a class="llms-button-text llms-collapse-toggle" data-action="close" href="#"><?php _e( 'Close All', 'lifterlms' ); ?></a>
+
+				<?php do_action( 'lifterlms_outline_after_footer' ); ?>
 
 			</li>
 
@@ -86,5 +100,6 @@
 
 	</ul>
 
+	<?php do_action( 'lifterlms_outline_after' ); ?>
 
 </div>